requirements.txt
DBUtils==1.3
Flask==0.12.2
Jinja2==2.9.6
PyMySQL==0.7.11
requests==2.18.3
websocket-client==0.56.0
